Fix changebar pdf build and add to Travis releases
Fix the changebar build caused by a bad interaction between Sphinx and
latexdiff. latexdiff was not doing the right thing for changes in
chapter headers due to "red" and "blue" getting uppercased to "RED" and
"BLUE", and creating invalid output when adding some verbatim sections.
Fixed by moving to latest latexdiff and adding sphinxVerbatim to the
VERBATIMENV config variable.

Also update the travis-ci control file to also build the changebar
version of the document.

Cc: Rob Herring <robh@kernel.org>
Signed-off-by: Grant Likely <grant.likely@arm.com>
Signed-off-by: Rob Herring <robh@kernel.org>